better performance

+++ lib/SormVersion.php
@@ -22,24 +22,20 @@ class SormVersion extends SimpleORMap {
             ");
             $statement->execute(array($deleting));
         }
-        while (Config::get()->DELOREAN_MAX_SIZE > 0 && self::getAllocatedSpace() > Config::get()->DELOREAN_MAX_SIZE) {
-            $last = self::findOneBySQL("1 = 1 ORDER BY version_id ASC LIMIT 1");
-            if ($last) {
-                $last->delete();
-            } else {
-                break;
-            }
-        }
-        //verwaiste Dateien löschen:
-        $folder = self::getFileDataPath();
-        if ($folder) {
-            $files = array_diff(scandir($folder), array('.', '..'));
-            foreach ($files as $file) {
-                if (file_exists($folder . "/" . $file)) {
-                    $version = self::findOneBySQL("file_id = ? LIMIT 1", array($file));
-                    if (!$version) {
-                        @unlink($folder . "/" . $file);
+
+        if (Config::get()->DELOREAN_MAX_SIZE > 0) {
+            $old_allocated_space = self::getAllocatedSpace();
+            $freeed_space = 0;
+            while ($old_allocated_space - $freeed_space > Config::get()->DELOREAN_MAX_SIZE) {
+                $last = self::findOneBySQL("1 = 1 ORDER BY version_id ASC LIMIT 1");
+                if ($last) {
+                    $freeed_space += strlen($last['json_data']) + 20 + 32 + 128 + 97 + 100 + 100 + 4 + 4 + 11;
+                    if (file_exists($last->getFilePath())) {
+                        $freeed_space += filesize($last->getFilePath());
                     }
+                    $last->delete();
+                } else {
+                    break;
                 }
             }
         }
@@ -148,10 +144,20 @@ class SormVersion extends SimpleORMap {
     }
 
     public function cbCleanUp() {
-        while (Config::get()->DELOREAN_MAX_SIZE > 0 && self::getAllocatedSpace() > Config::get()->DELOREAN_MAX_SIZE) {
-            $last = self::findOneBySQL("1 = 1 ORDER BY version_id ASC LIMIT 1");
-            if ($last) {
-                $last->delete();
+        if (Config::get()->DELOREAN_MAX_SIZE > 0) {
+            $old_allocated_space = self::getAllocatedSpace();
+            $freeed_space = 0;
+            while ($old_allocated_space - $freeed_space > Config::get()->DELOREAN_MAX_SIZE) {
+                $last = self::findOneBySQL("1 = 1 ORDER BY version_id ASC LIMIT 1");
+                if ($last) {
+                    $freeed_space += strlen($last['json_data']) + 20 + 32 + 128 + 97 + 100 + 100 + 4 + 4 + 11;
+                    if (file_exists($last->getFilePath())) {
+                        $freeed_space += filesize($last->getFilePath());
+                    }
+                    $last->delete();
+                } else {
+                    break;
+                }
             }
         }
     }
